requirements.txt
Flask
pymongo
firebase-admin
python-dotenv
pymongo[srv]
firebase_admin
